+++ crates/serialization/src/python/mod.rs
@@ -63,6 +63,10 @@ pub fn serialization(_: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
         crate::python::arrow::py_index_prices_to_arrow_record_batch_bytes,
         m
     )?)?;
+    m.add_function(wrap_pyfunction!(
+        crate::python::arrow::py_instrument_closes_to_arrow_record_batch_bytes,
+        m
+    )?)?;
 
     m.add_class::<crate::enums::ParquetWriteMode>()?;