+++ nautilus_trader/serialization/arrow/schema.py
@@ -105,6 +105,12 @@ NAUTILUS_ARROW_SCHEMA = {
         },
         metadata={"type": "InstrumentClose"},
     ),
+    # InstrumentClose: pa.schema(  # TODO: Not implemented yet
+    #     [
+    #         pa.field(k, infer_dtype(v), False)
+    #         for k, v in nautilus_pyo3.InstrumentClose.get_fields().items()
+    #     ],
+    # ),
     InstrumentStatus: pa.schema(
         {
             "instrument_id": pa.dictionary(pa.int64(), pa.string()),